chore: added measures toolbar and sidebar components
